backend/requirements.txt
django-ninja>=1.4.3,<2.0
gunicorn>=23.0.0,<24.0.0
uvicorn[standard]>=0.34.3,<1.0
psycopg2>=2.9.10,<3.0